requirements.txt
streamlit
torch
torchvision
pillow
numpy
opencv-python-headless
uvicorn
fastapi